# Optimize cloning for Access-related structs